12.4继续过了一部分测例，主要修正了ASTReader/ASTWriter部分

+++ clang/lib/AST/ExternalASTSource.cpp
@@ -72,7 +72,7 @@ Decl *ExternalASTSource::GetExternalDecl(uint32_t ID) {
   return nullptr;
 }
 
-Selector ExternalASTSource::GetExternalSelector(uint64_t ID) {
+Selector ExternalASTSource::GetExternalSelector(uint32_t ID) {
   return Selector();
 }
